fix: Change FFI pointer types from u32 to i32 for wasm32 compatibility
- Changed to_passback_ptr return type from u32 to i32
- Cast all to_ptr + 4 expressions to i32 for FFI calls
- Fixes type mismatches in __load_context, __load_block, __load_transaction, __balance, __sequence, __fuel, __height, and __log calls
- Ensures compatibility with WebAssembly i32 function signatures

+++ crates/metashrew-support/src/compat.rs
@@ -98,8 +98,8 @@ pub fn to_ptr(v: &mut Vec<u8>) -> u32 {
 /// - Bytes 4+: Actual data payload
 /// 
 /// This function returns a pointer to the data payload, skipping the length.
-pub fn to_passback_ptr(v: &mut Vec<u8>) -> u32 {
-    to_ptr(v) + 4
+pub fn to_passback_ptr(v: &mut Vec<u8>) -> i32 {
+    (to_ptr(v) + 4) as i32
 }
 
 /// Convert data to ArrayBuffer layout with length prefix.
